Add files via upload
In this study, we applied feature selection to identify the most important genes from the microarray dataset for diabetes detection. Initially, we selected the top 80% of features, which improved model performance by reducing noise and overfitting. Then, by narrowing the selection to the top 60% of features, the model's accuracy remained robust, though with slight performance drops. Finally, using the top 40% of features resulted in a more compact model, still performing adequately for diabetes detection, indicating that a smaller set of highly relevant genes can be sufficient. These results demonstrate the effectiveness of feature selection in improving classifier performance while simplifying the model.